+++ src/include/extern.h
@@ -22,7 +22,7 @@ extern int __wt_block_addr_string(WT_SESSION_IMPL *session,
 extern int __wt_block_alloc( WT_SESSION_IMPL *session,
     WT_BLOCK *block,
     off_t *offsetp,
-    uint32_t size);
+    off_t size);
 extern int __wt_block_free_buf(WT_SESSION_IMPL *session,
     WT_BLOCK *block,
     const uint8_t *addr,
@@ -30,7 +30,7 @@ extern int __wt_block_free_buf(WT_SESSION_IMPL *session,
 extern int __wt_block_free( WT_SESSION_IMPL *session,
     WT_BLOCK *block,
     off_t off,
-    uint32_t size);
+    off_t size);
 extern void __wt_block_freelist_open(WT_SESSION_IMPL *session, WT_BLOCK *block);
 extern int __wt_block_freelist_read(WT_SESSION_IMPL *session, WT_BLOCK *block);
 extern void __wt_block_freelist_close(WT_SESSION_IMPL *session,
